requirements.txt
bs4 >= 0.0.1
python-dotenv >= 0.19.0
requests >= 2.26.0